+++ src/MapHeatmap.js
