+++ src/MapPolygon.js
