+++ src/decorateMapComponent.js
